Score Count, Final Score and Bottom Wall added.
Incremented the player score every time a saw overlap() occurs in PlayState. Also created a bottom wall FlxObject that doesn't move to block player movement from going off the bottom screen.

There may be another way to do it, but the only way I was able to get the total/final score to show up on the GameOverState screen was to make the variable score in the Hud class "public static var score:Int".